Initial implementation of entrance exam scoring.
This is a temporary solution for entry work checking in year 2016. It
just shows the panel with scoring results if the module is present and
there is at least one scorer.

Intended to be removed after check in 2016 and replaced by more general
and sensible solution. The removal won't cause loosing of any valuble
data.