Futher split dates migration for entrance steps
The initially splitted migration was still failing with the same error.
andgein@ has a suggestion, that the reason is adding a column to schema
and then trying to modify data in another column (maybe also added). I
cannot test it in my local setup, so trying to do it in prod O_o